Improve nav
This generally reimplements the navigation bar.

The navigation bar still largely looks and is interacted with as
before; the change is mainly technical.

The nav now uses <a></a>s instead of <a><button></button></a>s.
This simplifies the code and makes the page more similar to how
the browser expects pages to be programmed, which in turn makes
browser features such as tab-focusing work better.